requirements/base.txt
Django==1.10.4
django-cms==3.4.1
djangocms-text-ckeditor==3.3.1
djangocms-picture==2.0.4
djangocms-vimeo==0.4.0
django-environ==0.4.0
django-filer==1.2.5
django-polymorphic==1.0.2
django-ckeditor==5.2.1
django-bootstrap3==7.1.0
requests==2.12.4
django-extensions==1.7.5
django-js-reverse==0.7.3
django-machina==0.5.0
Pillow==3.4.2
webcolors==1.7
django-settings-export==1.2.1
django-colorfield==0.1.11
singledispatch==3.4.0.3
django-compressor==2.1
slimit==0.8.1
csscompressor==0.9.4
django-ordered-model==1.3.0
django-el-pagination==3.0.1
Whoosh==2.7.4
Markdown==2.6.8
django-mobile==0.7.0
django-decorator-include==1.2
django-nocaptcha-recaptcha==0.0.19
